rag-mcp-server/requirements.txt
mcp-server>=0.1.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
supabase>=2.0.0
numpy>=1.24.0
openai>=1.0.0
